requirements.txt
streamlit
watchdog
pydantic-settings
pandas
numpy
requests
pandas
pyyaml
pytest
httpx